+++ benchmarks/internal/cluster/worker.js
@@ -1,10 +1,13 @@
 'use strict'
+const { isMaster } = require('cluster')
 const { ClusterWorker } = require('../../../lib/index')
 const { jsonIntegerSerialization } = require('../benchmark-utils')
 
+const debug = false
+
 function yourFunction (data) {
   jsonIntegerSerialization(1000)
-  // console.log('This is the main thread ' + isMaster)
+  debug === true && console.debug('This is the main thread ' + isMaster)
   return { ok: 1 }
 }